// generated/graphql-storefront.ts
export type Maybe<T> = T | null
// ...
export type Article = HasMetafields &
	Node &
	OnlineStorePublishable &
	Trackable & {
		__typename?: "Article"
		/**
		 * The article's author.
		 * @deprecated Use `authorV2` instead.
		 */
		author: ArticleAuthor
		/** The article's author. */
		authorV2?: Maybe<ArticleAuthor>
		/** The blog that the article belongs to. */
		blog: Blog
		/** List of comments posted on the article. */
		comments: CommentConnection
		/** Stripped content of the article, single line with HTML tags removed. */
		content: Scalars["String"]["output"]
		/** The content of the article, complete with HTML formatting. */
		contentHtml: Scalars["HTML"]["output"]
		/** Stripped excerpt of the article, single line with HTML tags removed. */
		excerpt?: Maybe<Scalars["String"]["output"]>
		/** The excerpt of the article, complete with HTML formatting. */
		excerptHtml?: Maybe<Scalars["HTML"]["output"]>
		/** A human-friendly unique string for the Article automatically generated from its title. */
		handle: Scalars["String"]["output"]
		/** A globally-unique ID. */
		id: Scalars["ID"]["output"]
		/** The image associated with the article. */
		image?: Maybe<Image>
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>
		/** The metafields associated with the resource matching the supplied list of namespaces and keys. */
		metafields: Array<Maybe<Metafield>>
		/** The URL used for viewing the resource on the shop's Online Store. Returns `null` if the resource is currently not published to the Online Store sales channel. */
		onlineStoreUrl?: Maybe<Scalars["URL"]["output"]>
		/** The date and time when the article was published. */
		publishedAt: Scalars["DateTime"]["output"]
		/** The article’s SEO information. */
		seo?: Maybe<Seo>
		/**
		 * A categorization that a article can be tagged with.
		 *
		 */
		tags: Array<Scalars["String"]["output"]>
		/** The article’s name. */
		title: Scalars["String"]["output"]
		/** A URL parameters to be added to a page URL when it is linked from a GraphQL result. This allows for tracking the origin of the traffic. */
		trackingParameters?: Maybe<Scalars["String"]["output"]>
	}
// ...
export type ArticleConnection = {
	__typename?: "ArticleConnection"
	/** A list of edges. */
	edges: Array<ArticleEdge>
	/** A list of the nodes contained in ArticleEdge. */
	nodes: Array<Article>
	/** Information to aid in pagination. */
	pageInfo: PageInfo
}
// ...
export type Blog = HasMetafields &
	Node &
	OnlineStorePublishable & {
		__typename?: "Blog"
		/** Find an article by its handle. */
		articleByHandle?: Maybe<Article>
		/** List of the blog's articles. */
		articles: ArticleConnection
		/** The authors who have contributed to the blog. */
		authors: Array<ArticleAuthor>
		/**
		 * A human-friendly unique string for the Blog automatically generated from its title.
		 *
		 */
		handle: Scalars["String"]["output"]
		/** A globally-unique ID. */
		id: Scalars["ID"]["output"]
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>
		/** The metafields associated with the resource matching the supplied list of namespaces and keys. */
		metafields: Array<Maybe<Metafield>>
		/** The URL used for viewing the resource on the shop's Online Store. Returns `null` if the resource is currently not published to the Online Store sales channel. */
		onlineStoreUrl?: Maybe<Scalars["URL"]["output"]>
		/** The blog's SEO information. */
		seo?: Maybe<Seo>
		/** The blogs’s title. */
		title: Scalars["String"]["output"]
	}
// ...
export type Cart = HasMetafields &
	Node & {
		__typename?: "Cart"
		/** An attribute associated with the cart. */
		attribute?: Maybe<Attribute>
		/** The attributes associated with the cart. Attributes are represented as key-value pairs. */
		attributes: Array<Attribute>
		/** Information about the buyer that's interacting with the cart. */
		buyerIdentity: CartBuyerIdentity
		/** The URL of the checkout for the cart. */
		checkoutUrl: Scalars["URL"]["output"]
		/** The estimated costs that the buyer will pay at checkout. The costs are subject to change and changes will be reflected at checkout. The `cost` field uses the `buyerIdentity` field to determine [international pricing](https://shopify.dev/custom-storefronts/internationalization/international-pricing). */
		cost: CartCost
		/** The date and time when the cart was created. */
		createdAt: Scalars["DateTime"]["output"]
		/**
		 * The delivery groups available for the cart, based on the buyer identity default
		 * delivery address preference or the default address of the logged-in customer.
		 *
		 */
		deliveryGroups: CartDeliveryGroupConnection
		/** The discounts that have been applied to the entire cart. */
		discountAllocations: Array<CartDiscountAllocation>
		/** The case-insensitive discount codes that the customer added at checkout. */
		discountCodes: Array<CartDiscountCode>
		/**
		 * The estimated costs that the buyer will pay at checkout. The estimated costs are subject to change and changes will be reflected at checkout. The `estimatedCost` field uses the `buyerIdentity` field to determine [international pricing](https://shopify.dev/custom-storefronts/internationalization/international-pricing).
		 * @deprecated Use `cost` instead.
		 */
		estimatedCost: CartEstimatedCost
		/** A globally-unique ID. */
		id: Scalars["ID"]["output"]
		/** A list of lines containing information about the items the customer intends to purchase. */
		lines: BaseCartLineConnection
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>
		/** The metafields associated with the resource matching the supplied list of namespaces and keys. */
		metafields: Array<Maybe<Metafield>>
		/** A note that's associated with the cart. For example, the note can be a personalized message to the buyer. */
		note?: Maybe<Scalars["String"]["output"]>
		/** The total number of items in the cart. */
		totalQuantity: Scalars["Int"]["output"]
		/** The date and time when the cart was updated. */
		updatedAt: Scalars["DateTime"]["output"]
	}
// ...
export type Location = HasMetafields &
	Node & {
		__typename?: "Location"
		/** The address of the location. */
		address: LocationAddress
		/** A globally-unique ID. */
		id: Scalars["ID"]["output"]
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>
		/** The metafields associated with the resource matching the supplied list of namespaces and keys. */
		metafields: Array<Maybe<Metafield>>
		/** The name of the location. */
		name: Scalars["String"]["output"]
	}
// ...
export type Market = HasMetafields &
	Node & {
		__typename?: "Market"
		/**
		 * A human-readable unique string for the market automatically generated from its title.
		 *
		 */
		handle: Scalars["String"]["output"]
		/** A globally-unique ID. */
		id: Scalars["ID"]["output"]
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>
		/** The metafields associated with the resource matching the supplied list of namespaces and keys. */
		metafields: Array<Maybe<Metafield>>
	}
// ...
export type Metafield = Node & {
	__typename?: "Metafield"
	/** The date and time when the storefront metafield was created. */
	createdAt: Scalars["DateTime"]["output"]
	/** The description of a metafield. */
	description?: Maybe<Scalars["String"]["output"]>
	/** A globally-unique ID. */
	id: Scalars["ID"]["output"]
	/** The unique identifier for the metafield within its namespace. */
	key: Scalars["String"]["output"]
	/** The container for a group of metafields that the metafield is associated with. */
	namespace: Scalars["String"]["output"]
	/** The type of resource that the metafield is attached to. */
	parentResource: MetafieldParentResource
	/** Returns a reference object if the metafield's type is a resource reference. */
	reference?: Maybe<MetafieldReference>
	/** A list of reference objects if the metafield's type is a resource reference list. */
	references?: Maybe<MetafieldReferenceConnection>
	/**
	 * The type name of the metafield.
	 * Refer to the list of [supported types](https://shopify.dev/apps/metafields/definitions/types).
	 *
	 */
	type: Scalars["String"]["output"]
	/** The date and time when the metafield was last updated. */
	updatedAt: Scalars["DateTime"]["output"]
	/** The data stored in the metafield. Always stored as a string, regardless of the metafield's type. */
	value: Scalars["String"]["output"]
}
// ...
export type Node = {
	/** A globally-unique ID. */
	id: Scalars["ID"]["output"]
}
// ...
export type StorefrontMetafieldReferencesArticlesQuery = {
	__typename?: "QueryRoot"
	records: {
		__typename?: "ArticleConnection"
		nodes: Array<{
			__typename?: "Article"
			handle: string
			id: string
			metafield?: {
				__typename?: "Metafield"
				id: string
				namespace: string
				key: string
				type: string
				value: string
				parentResource:
					| { __typename: "Article"; handle: string; id: string }
					| { __typename: "Blog"; handle: string; id: string }
					| { __typename: "Cart" }
					| { __typename: "Collection" }
					| { __typename: "Customer" }
					| { __typename: "Location" }
					| { __typename: "Market" }
					| { __typename: "Order" }
					| { __typename: "Page"; handle: string; id: string }
					| { __typename: "Product" }
					| { __typename: "ProductVariant" }
					| { __typename: "Shop" }
			} | null
		}>
		pageInfo: {
			__typename?: "PageInfo"
			endCursor?: string | null
			hasNextPage: boolean
		}
	}
}

export type StorefrontMetafieldReferencesBlogsQueryVariables = Exact<{
	namespace: Scalars["String"]["input"]
	key: Scalars["String"]["input"]
	first: Scalars["Int"]["input"]
	after?: InputMaybe<Scalars["String"]["input"]>
}>

export type StorefrontMetafieldReferencesBlogsQuery = {
	__typename?: "QueryRoot"
	records: {
		__typename?: "BlogConnection"
		nodes: Array<{
			__typename?: "Blog"
			handle: string
			id: string
			metafield?: {
				__typename?: "Metafield"
				id: string
				namespace: string
				key: string
				type: string
				value: string
				parentResource:
					| { __typename: "Article"; handle: string; id: string }
					| { __typename: "Blog"; handle: string; id: string }
					| { __typename: "Cart" }
					| { __typename: "Collection" }
					| { __typename: "Customer" }
					| { __typename: "Location" }
					| { __typename: "Market" }
					| { __typename: "Order" }
					| { __typename: "Page"; handle: string; id: string }
					| { __typename: "Product" }
					| { __typename: "ProductVariant" }
					| { __typename: "Shop" }
			} | null
		}>
		pageInfo: {
			__typename?: "PageInfo"
			endCursor?: string | null
			hasNextPage: boolean
		}
	}
}

export type StorefrontMetafieldReferencesPagesQueryVariables = Exact<{
	namespace: Scalars["String"]["input"]
	key: Scalars["String"]["input"]
	first: Scalars["Int"]["input"]
	after?: InputMaybe<Scalars["String"]["input"]>
}>

export type StorefrontMetafieldReferencesPagesQuery = {
	__typename?: "QueryRoot"
	records: {
		__typename?: "PageConnection"
		nodes: Array<{
			__typename?: "Page"
			handle: string
			id: string
			metafield?: {
				__typename?: "Metafield"
				id: string
				namespace: string
				key: string
				type: string
				value: string
				parentResource:
					| { __typename: "Article"; handle: string; id: string }
					| { __typename: "Blog"; handle: string; id: string }
					| { __typename: "Cart" }
					| { __typename: "Collection" }
					| { __typename: "Customer" }
					| { __typename: "Location" }
					| { __typename: "Market" }
					| { __typename: "Order" }
					| { __typename: "Page"; handle: string; id: string }
					| { __typename: "Product" }
					| { __typename: "ProductVariant" }
					| { __typename: "Shop" }
			} | null
		}>
		pageInfo: {
			__typename?: "PageInfo"
			endCursor?: string | null
			hasNextPage: boolean
		}
	}
}
// ...
export type StorefrontMetafieldFragmentFragment = {
	__typename?: "Metafield"
	id: string
	namespace: string
	key: string
	type: string
	value: string
	parentResource:
		| { __typename: "Article"; handle: string; id: string }
		| { __typename: "Blog"; handle: string; id: string }
		| { __typename: "Cart" }
		| { __typename: "Collection" }
		| { __typename: "Customer" }
		| { __typename: "Location" }
		| { __typename: "Market" }
		| { __typename: "Order" }
		| { __typename: "Page"; handle: string; id: string }
		| { __typename: "Product" }
		| { __typename: "ProductVariant" }
		| { __typename: "Shop" }
}